Replaced container subid allocator implementation.
This change replaces the subid allocator in Sysbox with a simple
implementation that assigns all containers the same subid range.

A more advanced implementation of the subid allocator is present
in sysbox enterprise (sysbox-ee).
